projects/test_work_fullstack/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
email-validator>=2.0.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
sqlalchemy>=2.0.36
alembic==1.13.1
psycopg[binary]>=3.2.0
